requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bcrypt==4.0.1
brainflow==0.0.1
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
cmake==4.2.1
contourpy==1.3.3
cryptography==46.0.4
cycler==0.12.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
filelock==3.24.3
flatbuffers==20181003210633
fonttools==4.61.1
fsspec==2026.2.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jax==0.7.1
jaxlib==0.7.1
Jinja2==3.1.6
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.18
ml_dtypes==0.5.4
mpmath==1.3.0
mysql-connector-python==9.6.0
networkx==3.6.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.8.0.76
opt_einsum==3.4.0
packaging==26.0
passlib==1.7.4
pillow==12.0.0
protobuf==4.25.8
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pyotp==2.9.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
qrcode==8.2
rich==14.3.1
rich-toolkit==0.17.2
rignore==0.7.6
rsa==4.9.1
scipy==1.17.0
sentencepiece==0.2.1
sentry-sdk==2.51.0
shellingham==1.5.4
six==1.17.0
sounddevice==0.5.5
starlette==0.50.0
sympy==1.14.0
tflite-runtime==2.14.0
torch==2.10.0+cpu
torchaudio==2.10.0+cpu
torchvision==0.25.0+cpu
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
